removed slack integration, removed ImageService, GitHubHookService, PublishWeekStatsService which were not used in the code
