+++ bot/src/main/java/com/community/tools/controller/GitUsersController.java
@@ -1,0 +1,42 @@
+package com.community.tools.controller;
+
+import com.community.tools.service.github.GitHubService;
+import io.swagger.annotations.ApiOperation;
+import lombok.RequiredArgsConstructor;
+import org.kohsuke.github.GHPerson;
+import org.kohsuke.github.GHUser;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import static org.springframework.http.ResponseEntity.ok;
+
+@RequiredArgsConstructor
+@RestController
+@RequestMapping("app")
+public class GitUsersController {
+  private final GitHubService gitService;
+
+  /**
+   * Endpoint /git. Method GET.
+   *
+   * @return ResponseEntity with Status.OK and List of all users in GH repository
+   */
+  @ApiOperation(value = "Returns list of github logins"
+      + " of Broscorp-net/traineeship collaborators")
+  @GetMapping(value = "/git", produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<List<String>> getGitHubAllUsers() {
+    Set<GHUser> gitHubAllUsers = gitService.getGitHubAllUsers();
+
+    List<String> listGitUsersLogin = gitHubAllUsers.stream().map(GHPerson::getLogin)
+        .collect(Collectors.toList());
+
+    return ok().body(listGitUsersLogin);
+  }
+}
