+++ bot/src/main/java/com/community/tools/service/github/GitHubService.java
@@ -36,34 +36,6 @@ public class GitHubService {
   private final GitHubConnectService service;
 
   /**
-   * Get GitHub pull requests according to state.
-   *
-   * @param statePullRequest state of pull. T - open, F - closed
-   * @return Map of GH login trainee as a key, title of pull as value
-   */
-  public Map<String, String> getPullRequests(boolean statePullRequest) {
-    Map<String, String> listUsers = new HashMap<>();
-    try {
-      GHRepository repository = service.getGitHubRepository();
-      List<GHPullRequest> pullRequests;
-      if (!statePullRequest) {
-        pullRequests = repository.getPullRequests(GHIssueState.CLOSED);
-      } else {
-        pullRequests = repository.getPullRequests(GHIssueState.OPEN);
-      }
-
-      for (GHPullRequest repo : pullRequests) {
-        String login = repo.getUser().getLogin();
-        String title = repo.getTitle();
-        listUsers.put(login, title);
-      }
-    } catch (IOException e) {
-      throw new RuntimeException(e);
-    }
-    return listUsers;
-  }
-
-  /**
    * Get all events by the date interval.
    *
    * @param startDate startDate
@@ -133,29 +105,6 @@ public class GitHubService {
     } catch (IOException e) {
       throw new RuntimeException(e);
     }
-  }
-
-  /**
-   * Get active GitHub user names.
-   *
-   * @param date date
-   * @return Set of GH User names
-   */
-  public Set<String> getActiveUsersFromGit(Date date) {
-    Set<String> names = new HashSet<>();
-    try {
-      GHRepository repository = service.getGitHubRepository();
-      List<GHPullRequest> pullRequests = repository.getPullRequests(GHIssueState.ALL);
-
-      for (GHPullRequest pr : pullRequests) {
-        if (pr.getCreatedAt().after(date)) {
-          names.add(pr.getUser().getLogin());
-        }
-      }
-    } catch (IOException ex) {
-      throw new RuntimeException(ex);
-    }
-    return names;
   }
 
 }
